refactor(rust-cli): isolate verification orchestration
